WIP: DAECompiler v2
This is a WIP branch for IPO-enabled DAECompiler v2.
The main objective here is to:
- Remove all dead code from the various iterations of DAECompiler
- Fully transition to IPO caches
- Switch to the new intrinsic design
- Refactor the code to make it easier to understand

The way I have gone about this is to delete most of DAECompiler,
and then put back the pieces one by one in the proper place, updating
things as I go along. The hope is that this way, most of the code
has been looked at to make sure it's still needed and sensible.

The current status of this branch is that (with very recent master)
it passes the minimal smoke tests in test/ipo. However, I have not
yet put back Diffractor integration of written codegen for jacobians.

As such, it should be considered a proof of concept for the new caching
setup, rather than a finished implementation. Putting it up now for
early preview.