Add thread-through state to all partitions (#36)
This changes the ABI of our scheduled functions to return tuples of
the requested equation residuals and some thread-through state (these
are basically the same as the optic residuals in Diffractor, but
I'm avoiding using the terms residuals, because that already has
three meanings in DAECompiler). This temporarily somewhat regresses
SICM quality, but we need to do some further work to properly integrate
SICM with our initialization system, and I'd like to wait with restoring
SICM quality until that is done.