Correct var ssa mapping (#53)
Unlike equations, which have a fairly unambiguous ssa mapping, it's
not entirely clear what we should be displaying here. For now, make
this display only callee internal variables, which is the obvious
thing to do. I think it would also make sense to indicate variables
there were originated by a particular callee, but possibly in a
different colors, but I'm puting on that for this PR, since it
needs additional changes in StateSelection.